redesigns spec entry
This will require users to enter specs a bit differently, since there is
a couple of problems with the way we're handling this at the moment:

Before
------
```
// Request body as JSON:

{ "spec": { "openapi": "3.0.0" } } // etc.
```

This will break under certain circumstances as param mapping will not
accept this in Phoenix (since server lists can be maps with multiple
keys). It also omits the whole "we know whether this is YAML or JSON at
runtime".

After
-----
```
// Request body as JSON:

{ "json": " {\"openapi\": \"3.0.0\"}" } // etc.
```

This allows for knowing what spec we're parsing and will also open up
other entrypoints more easily (e.g. fetching from a URL, uploading
directly etc.)